RM4936: move tracking number from base to stock
TrackingNumber and TrackingNumberConfiguration are now in module stock.
Their package, every service and views, and translations are updated.
