+++ axelor-base/src/main/resources/i18n/messages.csv
@@ -19,7 +19,6 @@
 "+33100000000",,,
 "- Succeeded :",,,
 "0",,,
-"0 --> infinite",,,
 "00",,,
 "01",,,
 "02",,,
@@ -164,7 +163,6 @@
 "Actions",,,
 "Activate bar code generation for products",,,
 "Activate sending email",,,
-"Active",,,
 "Active ?",,,
 "Active Team",,,
 "Active company",,,
@@ -214,16 +212,12 @@
 "Annually",,,
 "Anomalies",,,
 "Anomaly",,,
-"App",,,
 "App base",,,
 "Apparel",,,
 "Applicable dates",,,
 "Application",,,
-"Application Config",,,
 "Application start Date",,,
 "Apply on",,,
-"Apps",,,
-"Apps management",,,
 "April",,,
 "Archived ?",,,
 "Are you sure you want to start the import?",,,
@@ -239,8 +233,6 @@
 "Attributes",,,
 "August",,,
 "Authority Partner",,,
-"Auto sale tracking Nbr. order",,,
-"Auto select sale tracking Nbr. ?",,,
 "Availability",,,
 "Available",,,
 "BBAN",,,
@@ -293,7 +285,6 @@
 "Blockings",,,
 "Both currencies must be saved before currency rate apply",,,
 "Branch Identifier (Branch Code)",,,
-"Bulk install",,,
 "Business Party Prefix (Bank Code)",,,
 "Business Party Suffix (Location Code)",,,
 "Busy",,,
@@ -412,8 +403,6 @@
 "Configuration",,,
 "Configuration file",,,
 "Configurations",,,
-"Configure",,,
-"Configure:",,,
 "Confirm",,,
 "Connection successful",,,
 "Contact",,,
@@ -428,7 +417,6 @@
 "Cost price",,,
 "Cost type",,,
 "Costs",,,
-"Counter",,,
 "Countries",,,
 "Country",,,
 "Country Code",,,
@@ -488,8 +476,6 @@
 "Delivery address",,,
 "Delivery delay (days)",,,
 "Delivery stock move",,,
-"Demo data loaded ?",,,
-"Demo data loaded sucessefully",,,
 "Demography",,,
 "Departement",,,
 "Department",,,
@@ -497,7 +483,6 @@
 "Department Name",,,
 "Department city List",,,
 "Departments",,,
-"Depends on",,,
 "Dept./Div.",,,
 "Description",,,
 "Destination Currency",,,
@@ -564,6 +549,7 @@
 "Fax",,,
 "Feature currently not available with Open Street Maps.",,,
 "February",,,
+"Fetch emails",,,
 "Field",,,
 "Fields",,,
 "File",,,
@@ -580,7 +566,6 @@
 "Fixed",,,
 "Fixed Amount/%",,,
 "Fixed phone",,,
-"Following apps will be installed %s",,,
 "Food & Beverage",,,
 "Format",,,
 "Formula",,,
@@ -601,9 +586,6 @@
 "General data",,,
 "General discount (%)",,,
 "Generate",,,
-"Generate new production auto tracking Nbr. ?",,,
-"Generate new purchase auto tracking Nbr. ?",,,
-"Generate new sale auto tracking Nbr. ?",,,
 "Generate product variant wizard",,,
 "Generate product variants",,,
 "Generated Project",,,
@@ -631,13 +613,10 @@
 "INSEE Code",,,
 "IPO",,,
 "IPO + Cheque",,,
-"Image",,,
 "Import",,,
 "Import Calendar",,,
 "Import calendar",,,
 "Import configuration",,,
-"Import demo data",,,
-"Import demo data ?",,,
 "Import identifier",,,
 "Import type",,,
 "Importer",,,
@@ -663,8 +642,6 @@
 "Industry sectors",,,
 "Information",,,
 "Informations",,,
-"Init data loaded ?",,,
-"Install",,,
 "Interfaces",,,
 "Internal companies",,,
 "Internal description",,,
@@ -726,9 +703,6 @@
 "Mail batches",,,
 "Main company",,,
 "Manage mail account by user ?",,,
-"Manage production tracking Nbr. ?",,,
-"Manage purchase tracking Nbr. ?",,,
-"Manage sale tracking Nbr. ?",,,
 "Management coef.",,,
 "Manager Type",,,
 "Manufacturing",,,
@@ -748,7 +722,6 @@
 "Mobile phone",,,
 "Model",,,
 "Modules configurations",,,
-"Modules contains in the app",,,
 "Monday",,,
 "Month",,,
 "Month/days ?",,,
@@ -793,8 +766,6 @@
 "Next theoretical execution date",,,
 "No",,,
 "No Discount",,,
-"No application language set. Please set 'application.locale' property.",,,
-"No configuraiton required",,,
 "No contact selected/created, do you want to create a new one?",,,
 "No currency conversion found from '%s' to '%s' for date %s",,,
 "No duplicate records found",,,
@@ -876,7 +847,6 @@
 "Period duration",,,
 "Periods list",,,
 "Perishable ?",,,
-"Perishable expiration date",,,
 "Phone Book",,,
 "PhoneBook",,,
 "Phonebook.phonebook",,,
@@ -969,13 +939,11 @@
 "ProductSheet.price",,,
 "ProductSheet.title",,,
 "ProductSheet.warranty",,,
-"Production Qty by tracking",,,
 "Production config",,,
 "Production config (${ name })",,,
 "Production informations",,,
 "Production order",,,
 "Production processes",,,
-"Production tracking management",,,
 "Products",,,
 "Products (${ name })",,,
 "Products ATI/WT",,,
@@ -999,14 +967,12 @@
 "Purchase / Cost currency",,,
 "Purchase Order",,,
 "Purchase Price List",,,
-"Purchase Qty by tracking",,,
 "Purchase Tax",,,
 "Purchase config",,,
 "Purchase config (${ name })",,,
 "Purchase interface",,,
 "Purchase order",,,
 "Purchase price W.T.",,,
-"Purchase tracking management",,,
 "Purchase/Costs",,,
 "Purchases",,,
 "QAS Moniker",,,
@@ -1027,7 +993,6 @@
 "Rate Date",,,
 "Rate updation",,,
 "Raw Materials",,,
-"Raw material",,,
 "Real Estate",,,
 "Real qty",,,
 "Receipt N°(Payment voucher)",,,
@@ -1088,7 +1053,6 @@
 "Sale Order",,,
 "Sale Price (for periodicity selected)",,,
 "Sale Price List",,,
-"Sale Qty by tracking",,,
 "Sale Tax",,,
 "Sale config",,,
 "Sale config (${ name })",,,
@@ -1097,7 +1061,6 @@
 "Sale price",,,
 "Sale price W.T.",,,
 "Sale supply default method",,,
-"Sale tracking management",,,
 "Sales",,,
 "Sales Manager",,,
 "Sales Turnover",,,
@@ -1115,7 +1078,6 @@
 "Sector",,,
 "Sector Keyword",,,
 "Sector Keywords",,,
-"Select all",,,
 "Sellable",,,
 "Send",,,
 "Send Email",,,
@@ -1216,11 +1178,9 @@
 "The following address is invalid. Please use '+' button to create a new one :",,,
 "The line cannot be imported (import : %s)",,,
 "The product pulled-off date must be greater than the launch date.",,,
-"The quantity by lot must be greater than 0",,,
 "The tags below can be used as a prefix or a suffix :",,,
 "There is no matching address in the QAS base",,,
 "There is no sequence set for the partners",,,
-"This app is used by %s. Please deactivate them before continue.",,,
 "This batch is not runnable !",,,
 "Thursday",,,
 "Ticket",,,
@@ -1241,12 +1201,7 @@
 "Tools",,,
 "Total variants",,,
 "Tracebacks",,,
-"Tracking Nbr.",,,
-"Tracking Nbr. Config.",,,
-"Tracking Number",,,
-"Tracking Number Configuration",,,
 "Tracking number",,,
-"Tracking number configuration",,,
 "Transfer",,,
 "Transport & Logistics",,,
 "Treasury interface",,,
@@ -1277,7 +1232,6 @@
 "Users",,,
 "Users list",,,
 "Users/Companies",,,
-"Users/Employees",,,
 "Valid query.",,,
 "Validate",,,
 "Validated",,,
@@ -1297,7 +1251,6 @@
 "WARNING : To Date must be after or equals to From Date",,,
 "WARNING : please close the current conversion period to create a new one.",,,
 "Warranty ?",,,
-"Warranty expiration date",,,
 "Warranty length (in months)",,,
 "Web",,,
 "Webservices",,,
@@ -1408,7 +1361,6 @@
 "help:Product.saleSupplySelect",,,
 "help:Product.startDate",,,
 "help:Product.supplierCatalogList",,,
-"help:Product.trackingNumberConfiguration",,,
 "help:Product.unit",,,
 "help:ProductCategory.parentProductCategory",,,
 "help:ProductCategory.productFamily",,,
