+++ axelor-base/src/main/resources/views/Menu.xml
@@ -259,14 +259,6 @@
 		<context name="fromMenu" expr="eval:true"/>		
 	</action-view>
 
-	<menuitem name="referential-conf-tracking-number-config" parent="referential-conf"
-		title="Tracking number configuration" action="referential.conf.tracking.number.config" />
-
-	<action-view name="referential.conf.tracking.number.config" title="Tracking number configuration"
-		model="com.axelor.apps.base.db.TrackingNumberConfiguration" >
-		<view type="grid" name="tracking-number-configuration-grid"/>
-		<view type="form" name="tracking-number-configuration-form"/>
-	</action-view>
 	
 	<menuitem name="referential-report" parent="referential-root" order="90"
 		title="Reportings"/>
