+++ axelor-stock/src/main/resources/domains/InventoryLine.xml
@@ -13,7 +13,7 @@
 	<decimal name="realQty" title="Real Qty"/>
 	<string name="description" title="Description"/>
 	
-	<many-to-one name="trackingNumber" ref="com.axelor.apps.base.db.TrackingNumber" title="TrackingNumber"/>
+	<many-to-one name="trackingNumber" ref="com.axelor.apps.stock.db.TrackingNumber" title="TrackingNumber"/>
     
   </entity>
 </domain-models>
