+++ axelor-stock/src/main/resources/domains/Product.xml
@@ -5,6 +5,7 @@
  
   <entity name="Product" lang="java">
       <decimal name="avgPrice" readonly="true" precision="20" scale="10" title="Average Price"/>
+      <many-to-one name="trackingNumberConfiguration" ref="com.axelor.apps.stock.db.TrackingNumberConfiguration" title="Tracking Nbr. Config." help="true"/>
   </entity>
 
 </domain-models>
