+++ axelor-stock/src/main/resources/domains/StockMoveLine.xml
@@ -16,7 +16,7 @@
 	<decimal name="realQty" title="Real Qty"/>
     <many-to-one name="unit" ref="com.axelor.apps.base.db.Unit" title="Unit"/>
     
-    <many-to-one name="trackingNumber" ref="com.axelor.apps.base.db.TrackingNumber" title="Tracking Nbr."/>
+    <many-to-one name="trackingNumber" ref="com.axelor.apps.stock.db.TrackingNumber" title="Tracking Nbr."/>
     
     <integer name="conformitySelect" title="Conformity" selection="stock.move.line.conformity.select"/>
