+++ axelor-stock/src/main/resources/domains/TrackingNumberConfiguration.xml
@@ -3,7 +3,7 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://axelor.com/xml/ns/domain-models http://axelor.com/xml/ns/domain-models/domain-models_5.0.xsd">
 
-  <module name="base" package="com.axelor.apps.base.db"/>
+  <module name="stock" package="com.axelor.apps.stock.db"/>
 
   <entity name="TrackingNumberConfiguration" lang="java">
