+++ axelor-stock/src/main/resources/i18n/messages_en.csv
@@ -1,6 +1,7 @@
 "key","message","comment","context"
 "${fullName} plan. st. move",,,
 "${fullName} real st. move",,,
+"0 --> infinite",,,
 "<B>%s or %s</B> not found",,,
 "A partial stock move has been generated (%s)",,,
 "Actions",,,
@@ -14,6 +15,8 @@
 "App stock",,,
 "Are you sure you want to realize this stock move?",,,
 "Arrivals History",,,
+"Auto sale tracking Nbr. order",,,
+"Auto select sale tracking Nbr. ?",,,
 "Average Price",,,
 "Basic",,,
 "Both source location address and delivery (to) address are required to define the delivery itinerary",,,
@@ -28,6 +31,7 @@
 "Configuration",,,
 "Conformity",,,
 "Consumed quantity",,,
+"Counter",,,
 "Country",,,
 "Create Stock Move",,,
 "Create inventory",,,
@@ -86,6 +90,9 @@
 "Generate inc. moves for non storable products ?",,,
 "Generate inc. moves for storable products ?",,,
 "Generate invoice",,,
+"Generate new production auto tracking Nbr. ?",,,
+"Generate new purchase auto tracking Nbr. ?",,,
+"Generate new sale auto tracking Nbr. ?",,,
 "Generate out. moves for non storable products ?",,,
 "Generate out. moves for storable products ?",,,
 "Generate reversion",,,
@@ -140,6 +147,9 @@
 "Location value",,,
 "Location(s)",,,
 "Manage BackOrder",,,
+"Manage production tracking Nbr. ?",,,
+"Manage purchase tracking Nbr. ?",,,
+"Manage sale tracking Nbr. ?",,,
 "Map",,,
 "Min. Qty",,,
 "Minimum stock rule",,,
@@ -164,6 +174,7 @@
 "Parent",,,
 "Partial stock move (From",,,
 "Partner",,,
+"Perishable expiration date",,,
 "Plan",,,
 "Planned",,,
 "Planned Inventories",,,
@@ -184,6 +195,7 @@
 "Product soon out of stock by location",,,
 "Product's stocks %s (%s) are not in sufficient quantity to realize the delivery",,,
 "Product's stocks %s (%s), tracking number {} are not in sufficient quantity to realize the delivery",,,
+"Production Qty by tracking",,,
 "Production Stock Move in Planned",,,
 "Production Stock Move out Planned",,,
 "Production forecasts",,,
@@ -191,7 +203,10 @@
 "Production order",,,
 "Production stock move in realized",,,
 "Production stock move out realized",,,
+"Production tracking management",,,
+"Purchase Qty by tracking",,,
 "Purchase order",,,
+"Purchase tracking management",,,
 "Quantity",,,
 "Real Qty",,,
 "Realize",,,
@@ -203,6 +218,9 @@
 "Reportings",,,
 "Return surplus",,,
 "Reverse stock move (From",,,
+"Sale Qty by tracking",,,
+"Sale tracking management",,,
+"Sequence",,,
 "Ship/Receive all products",,,
 "Shipment Mode",,,
 "Shipment Mode form",,,
@@ -288,6 +306,7 @@
 "Supplier virtual location",,,
 "Team",,,
 "The move doesn't contain any product, do you wish to continue ?",,,
+"The quantity by lot must be greater than 0",,,
 "There is currently no such file in the specified folder or the folder may not exists.",,,
 "There's already an existing storage, you must deactivate it first",,,
 "There's no configured sequence for inventory for company",,,
@@ -307,8 +326,11 @@
 "Total W.T.",,,
 "Tracking Nbr",,,
 "Tracking Nbr.",,,
+"Tracking Nbr. Config.",,,
 "Tracking Number",,,
+"Tracking Number Configuration",,,
 "Tracking Numbers",,,
+"Tracking number configuration",,,
 "TrackingNumber",,,
 "Type",,,
 "Type Status",,,
@@ -327,8 +349,10 @@
 "View Supplier deliveries",,,
 "Virtual",,,
 "Virtual locations informations",,,
+"Warranty expiration date",,,
 "You must configure a Stock module for the company %s",,,
 "You must configure a customer virtual location for the company %s",,,
 "You must configure a supplier virtual location for the company %s",,,
 "You must configure an inventory virtual location for the company %s",,,
 "You must select a stock location",,,
+"help:Product.trackingNumberConfiguration",,,
