+++ axelor-stock/src/main/resources/i18n/messages_fr.csv
@@ -1,6 +1,7 @@
 "key","message","comment","context"
 "${fullName} plan. st. move",,,
 "${fullName} real st. move",,,
+"0 --> infinite","0 --> infini",,
 "<B>%s or %s</B> not found","<B>%s ou %s</B> introuvables",,
 "A partial stock move has been generated (%s)","Un mouvement de stock partiel a été généré (%s)",,
 "Actions","Actions",,
@@ -14,6 +15,8 @@
 "App stock",,,
 "Are you sure you want to realize this stock move?","Confirmer la réalisation de ce mouvement de stock?",,
 "Arrivals History","Historique réceptions",,
+"Auto sale tracking Nbr. order","Sélection automatique numéro de série ventes",,
+"Auto select sale tracking Nbr. ?","Sélection automatique numéro de série ventes",,
 "Average Price","Prix Moyen",,
 "Basic","Basique",,
 "Both source location address and delivery (to) address are required to define the delivery itinerary","Les champs Emplacement source et A l'adresse sont requis pour définir l'itinéraire de livraison.",,
@@ -28,6 +31,7 @@
 "Configuration","Configuration",,
 "Conformity","Conformité",,
 "Consumed quantity","Quantité consommée",,
+"Counter","Compteur",,
 "Country","Pays",,
 "Create Stock Move","Créer Mvt. Stock",,
 "Create inventory","Créer inventaire",,
@@ -86,6 +90,9 @@
 "Generate inc. moves for non storable products ?","Générer Mvt. Réception pour les produits non stockables ? ",,
 "Generate inc. moves for storable products ?","Générer Mvt. Réception pour les produits stockables ? ",,
 "Generate invoice","Générer facture",,
+"Generate new production auto tracking Nbr. ?","Générer N° suivi auto pour production ?",,
+"Generate new purchase auto tracking Nbr. ?","Générer N° suivi auto pour achats ?",,
+"Generate new sale auto tracking Nbr. ?","Générer N° suivi auto pour ventes ?",,
 "Generate out. moves for non storable products ?","Générer Mvt. Livraison pour les produits non stockables ? ",,
 "Generate out. moves for storable products ?","Générer Mvt. Livraison pour les produits stockables ? ",,
 "Generate reversion","Réexpédier",,
@@ -140,6 +147,9 @@
 "Location value","Valeur emplacement",,
 "Location(s)",,,
 "Manage BackOrder","Gérer les reliquats",,
+"Manage production tracking Nbr. ?","Gérer N° de suivi de production ? ",,
+"Manage purchase tracking Nbr. ?","Gérer N° suivi à l'achat ?",,
+"Manage sale tracking Nbr. ?","Gérer N° suivi à la vente ? ",,
 "Map",,,
 "Min. Qty","Qté Min.",,
 "Minimum stock rule","Les Règles de stock minimum",,
@@ -164,6 +174,7 @@
 "Parent","Emplacement Parent",,
 "Partial stock move (From","Reliquat (Du",,
 "Partner","Tiers",,
+"Perishable expiration date","Date de péremption",,
 "Plan","Planifier",,
 "Planned","Planifié",,
 "Planned Inventories","Inventaires Planifiés",,
@@ -184,6 +195,7 @@
 "Product soon out of stock by location","Produits prochainement sans stock par emplacement",,
 "Product's stocks %s (%s) are not in sufficient quantity to realize the delivery","Les stocks du produit %s (%s) sont insuffisants pour réaliser la livraison",,
 "Product's stocks %s (%s), tracking number {} are not in sufficient quantity to realize the delivery","Les stocks du produit %s (%s), numéro de suivi {}  sont insuffisants pour réaliser la livraison",,
+"Production Qty by tracking","Qté par lot de production",,
 "Production Stock Move in Planned","Mvt. D'entrée en production planifié",,
 "Production Stock Move out Planned","Mvt. De sortie de production planifié",,
 "Production forecasts","Mvt. Prévisionnels de production",,
@@ -191,7 +203,10 @@
 "Production order","Ordre de production",,
 "Production stock move in realized","Mvt. d'entrée en production réalisé",,
 "Production stock move out realized","Mvt. De sortie de production réalisé",,
+"Production tracking management","Gestion lots de production",,
+"Purchase Qty by tracking","Qté par lot à l'achat",,
 "Purchase order","Devis fournisseurs",,
+"Purchase tracking management","Suivi Achats",,
 "Quantity","Quantité",,
 "Real Qty","Qté réelle",,
 "Realize","Réalisé",,
@@ -203,6 +218,9 @@
 "Reportings",,,
 "Return surplus","Générer les retours",,
 "Reverse stock move (From","Inverser mouvement de stock (Du",,
+"Sale Qty by tracking","Qté par lot à la vente",,
+"Sale tracking management","Suivi à la vente",,
+"Sequence",,,
 "Ship/Receive all products","Envois/Réception de produits",,
 "Shipment Mode","Mode de livraison",,
 "Shipment Mode form","Mode de livraison",,
@@ -288,6 +306,7 @@
 "Supplier virtual location","Emplacement virtuel fournisseur",,
 "Team","Equipe",,
 "The move doesn't contain any product, do you wish to continue ?","Le mouvement de stock ne contient aucun produit, souhaitez-vous continuer ? ",,
+"The quantity by lot must be greater than 0","La quantité par lot doit être supérieure à 0.",,
 "There is currently no such file in the specified folder or the folder may not exists.","Il n'y a actuelleement aucun fichier dans le dossier spécifique ou le dossier n'existe peut-être pas",,
 "There's already an existing storage, you must deactivate it first","Il existe déjà un entrepot par défaut, veuillez d'abord désactiver l'entrepot",,
 "There's no configured sequence for inventory for company","Aucune séquence configurée pour les inventaires pour la société",,
@@ -307,9 +326,12 @@
 "Total W.T.",,,
 "Tracking Nbr","N° Suivi/lot",,
 "Tracking Nbr.","N° Suivi/lot",,
+"Tracking Nbr. Config.","Config. N° Suivi/lots",,
 "Tracking Number","Numéro de tracking",,
-"Tracking Numbers","N° de suivis/lots",,
-"TrackingNumber",,,
+"Tracking Number Configuration","Configuration N° suivi/lots",,
+"Tracking Numbers","N° de suivi/lots",,
+"Tracking number configuration","Configuration N° suivi/lots",,
+"TrackingNumber","N° Suivi",,
 "Type","Type",,
 "Type Status","Type Statut",,
 "Unit","Unité",,
@@ -327,8 +349,10 @@
 "View Supplier deliveries","Voir Réceptions fournisseur",,
 "Virtual","Virtuel",,
 "Virtual locations informations","Informations Stocks Virtuels",,
+"Warranty expiration date","Date de fin de garantie",,
 "You must configure a Stock module for the company %s","Veuillez configurer le module stock pour la société %s",,
 "You must configure a customer virtual location for the company %s","Veuillez configurer un Emplacement Virtuel Client pour la société %s",,
 "You must configure a supplier virtual location for the company %s","Veuillez configurer un Emplacement Virtuel Fournisseur pour la société %s",,
 "You must configure an inventory virtual location for the company %s","Veuillez configurer un Emplacement Virtuel Inventaire pour la société %s",,
 "You must select a stock location","Veuillez selectionner un entrepot",,
+"help:Product.trackingNumberConfiguration",,,
